scripts/requirements.txt
pushbullet.py==0.11.0